requirements.txt
google-generativeai
python-dotenv
pydantic
rich
colorama
